Fixed camera perspective

+++ MintyCore/Systems/Client/ApplyGPUCameraBufferSystem.cs
@@ -42,7 +42,7 @@ namespace MintyCore.Systems.Client
 
 
 				var cameraMatrix = Matrix4x4.CreateLookAt(position.Value, position.Value + new Vector3(0, 0, -1), new Vector3(0, 1, 0));
-				var camProjection = Matrix4x4.CreatePerspectiveFieldOfView(camera.Fov, MintyCore.Window.GetWindow().Width / MintyCore.Window.GetWindow().Height, 0.1f, 200f);
+				var camProjection = Matrix4x4.CreatePerspectiveFieldOfView(camera.Fov, (float)MintyCore.Window.GetWindow().Width / MintyCore.Window.GetWindow().Height, 0.1f, 200f);
 				VulkanEngine.UpdateBuffer( _cameraBuffers[World][_frameNumber[World]].buffer , cameraMatrix * camProjection);
 			}
 		}
